# docs/swagger.yaml
definitions:
  controller.HomeData:
    properties:
      checkIn:
        type: boolean
      msgQuantity:
        type: integer
      notiNew:
        type: integer
      policyLink:
        type: string
      storyList:
        items:
          $ref: '#/definitions/protocol.Pre7Story'
        type: array
      termsLink:
        type: string
      videoChatList:
        items:
          $ref: '#/definitions/protocol.WTRoomUser'
        type: array
      voiceChatList:
        items:
          $ref: '#/definitions/protocol.WTRoomUser'
        type: array
    type: object
  protocol.Announcement:
    properties:
      an_body:
        type: string
      an_title:
        type: string
      an_url:
        type: string
      at_msg:
        type: string
      idx:
        type: integer
      stat:
        type: integer
    type: object
...
  protocol.Noti:
    properties:
      at_noti:
        type: string
      frm_nick:
        type: string
      frm_uid:
        type: integer
      frm_url:
        type: string
      idx:
        type: integer
      nt_msg:
        type: string
      nt_title:
        type: string
      nt_type:
        type: integer
      stat:
        type: integer
    type: object
...
  protocol.Pre7Story:
    properties:
      idx:
        type: integer
      nick:
        type: string
      pic1:
        type: string
    type: object
...
  protocol.RespHeader:
    properties:
      desc:
        type: string
      result:
        $ref: '#/definitions/protocol.ResultCode'
      resultString:
        type: string
    type: object
  protocol.ResultCode:
    enum:
    - 0
    type: integer
    x-enum-varnames:
    - Success
...
  protocol.WTRoomUser:
    properties:
      age:
        type: string
      area:
        type: string
      did:
        type: string
      gender:
        type: string
      intro:
        type: string
      mainPic:
        type: string
      newStat:
        type: boolean
      nick:
        type: string
      sid:
        type: string
      thumbPic:
        type: string
      uid:
        type: integer
    type: object
...
info:
  contact: {}
  description: ms-gateway API server
  title: ms-gateway API
  version: "1.0"
paths:
...
  /home/v01/mdata:
    get:
      consumes:
      - application/json
      description: Retrieves main information to be displayed on the home screen (new
        notifications, message count, check-in status, video/voice chat lists, story
        list, terms links)
      produces:
      - application/json
      responses:
        "200":
          description: Home screen information
          schema:
            $ref: '#/definitions/controller.HomeData'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/protocol.RespHeader'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/protocol.RespHeader'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/protocol.RespHeader'
      security:
      - JwtAuth: []
      summary: Retrieves main information to be displayed on the home screen
      tags:
      - Home
...
  /noti/v01/anc/detail/{idx}:
    get:
      consumes:
      - application/json
      description: 공지사항 상세 정보를 조회합니다
      parameters:
      - description: Announcement index
        in: path
        name: idx
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/protocol.Announcement'
        "400":
          description: idx is required
          schema:
            $ref: '#/definitions/protocol.RespHeader'
        "404":
          description: Announcement not found
          schema:
            $ref: '#/definitions/protocol.RespHeader'
        "500":
          description: Failed to get announcement detail
          schema:
            $ref: '#/definitions/protocol.RespHeader'
      summary: 공지사항 상세 조회
      tags:
      - Noti
  /noti/v01/anc/list:
    get:
      consumes:
      - application/json
      description: 공지사항 목록을 조회합니다 (일주일 이내 공지는 new 표시)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/protocol.Announcement'
            type: array
        "500":
          description: Failed to get announcement list
          schema:
            $ref: '#/definitions/protocol.RespHeader'
      summary: 공지사항 목록 조회
      tags:
      - Noti
  /noti/v01/detail/{idx}:
    get:
      consumes:
      - application/json
      description: 알림 상세 정보를 조회하고 읽음 처리합니다
      parameters:
      - description: Noti index
        in: path
        name: idx
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/protocol.Noti'
        "400":
          description: idx is required
          schema:
            $ref: '#/definitions/protocol.RespHeader'
        "500":
          description: Failed to get noti detail
          schema:
            $ref: '#/definitions/protocol.RespHeader'
      summary: 알림 상세 조회
      tags:
      - Noti
  /noti/v01/list:
    get:
      consumes:
      - application/json
      description: 사용자의 알림 목록을 조회합니다 (JWT 인증 기반)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/protocol.RespHeader'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/protocol.RespHeader'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/protocol.RespHeader'
      security:
      - BearerAuth: []
      summary: 알림 목록 조회
      tags:
      - Noti
...
  /user/v01/getset/{uid}:
    get:
      consumes:
      - application/json
      description: |-
        Retrieves notification settings for authenticated users via JWT token.
        Notification settings are managed using bit flags:
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notification setting value (0-15)
          schema:
            allOf:
            - $ref: '#/definitions/protocol.RespHeader'
            - properties:
                data:
                  type: integer
              type: object
        "401":
          description: Authentication failed - No JWT token
          schema:
            $ref: '#/definitions/protocol.RespHeader'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/protocol.RespHeader'
      security:
      - BearerAuth: []
      summary: Get user notification settings
      tags:
      - Setting
  /user/v01/set/{cate}/{value}:
    post:
      consumes:
      - application/json
      description: |-
        Updates user settings based on category. Currently supports notification alert settings.
        Alert settings use bit flags (0-15) with XOR operation to toggle:
        - noti=1 (notification), call=2 (call), msg=4 (message), rvr=8 (reserve)
        - 0 = all alerts enabled
        - 15 = all alerts disabled
        - Example: To toggle notification, send value "1"
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Setting category (e.g. alert)
        in: path
        name: cate
        required: true
        type: string
      - description: Setting value
        in: path
        name: value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Setting updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request - Missing or invalid parameters
          schema:
            $ref: '#/definitions/protocol.RespHeader'
        "401":
          description: Authentication failed - No JWT token
          schema:
            $ref: '#/definitions/protocol.RespHeader'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/protocol.RespHeader'
      security:
      - BearerAuth: []
      summary: Update user settings
      tags:
      - Setting
...
swagger: "2.0"